pyproject.toml
[project]
name = "vtt"
version = "0.1.0"
description = "VTT(Video To Telegram), the opensource package for generating summary via opanAI from youtube daily then send to telegram."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.188.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
    "openai>=2.14.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.6",
    "pytube>=15.0.0",
    "youtube-transcript-api>=1.2.4",
    "yt-dlp>=2026.2.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vtt"]


